move tests to `webapis_tests.js`
